Refactor enemy and player classes to entities folder, create boss file
